Found a fast point searcher!
The previous strategy added just e_i - e_{i+1} to existing points.
The new strategy also notes which points are good "translators"
(i.e. are close to the rendered plane) and adds them to the list of
"directions" (i.e. things to try adding to exisitng points).

This also includes a slightly modified translator search strategy.

Work that remains in this area includes:

* Tweak the functions giving various measures of suitability:
    * as translators = as search directions, assumes this update
    * as the translator to use before drawing the current frame
    * as border points from which to continue searching
* Decide on a strategy for halting the search.
    * Resuming may not be necessary if the search is fast enough.