Move mouse motion to motion.js.
Now there's a default implementation of Movable.moveFromTo() for mouse
motion. Also implemented a crude sort of mouse coasting.

To ease this, decided to use an upper-left origin consistently
throughout Movable, and perform a scaling before using the value of
Movable.position.

Also wrote a replaceLattice() function in preparation for hot-swappable
QuasiLattice objects.