Implement a QuasiLattice base class.
The representation is now one of the inputs.

Now the 3D quasi-lattice inherits from the base class.
Still remaining is to bring the 2D version into this framework.

In theory it should be easier to swap in the BC_3 Coxeter
system for the 3D version now.

And someday it'd be nice to have all the information stored
outside the QL and Vertex classes so that I only need to make
subclasses to assist with presentation.