2D: Zooming and reduction-to-totient.
Permanently breaks two-canvas rendering.

By "reduction-to-totient" I mean restricting to the faithful
subrepresentation (over Z) of the Cyclic(n) (whose dimension is
the totient of n). This makes basically the same patterns---the
two visible dimensions are drawn from the faithful part,
and it was necessary to make the algebra work out for zooming.

Two-canvas rendering was (is likely still) a neat idea, but my
implementation suffered from flicker/stutter and made it much
harder for me to figure out how to modify the drawing process.
So in order to implmement zooming I trashed it. Maybe later I'll
try for a less intrusive implementation.

In the process I also imported a bunch of logic from the 3D case:
* Touch interface and zooming calculations
* When to stop asking for more verts
...likely with more stuff I've forgotten but intend to unify where
it makes sense.

Zooming requires choosing a real unit from the algebraic integers
of the nth cyclotomic field. Ideally for zooming purposes only one
of the real embeddings would put it outside of (-1,1), but I don't
know how to choose a unit with that property---especially not if I
also want the one place outsie (-1,1) to be close to 1.