2D: Smarter unit choice.
Now the unit chosen will give shrinking (still non-uniform except
for n = 5) in all unseen dimensions up through n = 10.
After that, at least one unseen dimension becomes growing,
except in n = 12, 14, 15, 18, 24, ...

I think that's the complete list, actually; but it doesn't seem
terribly interesting to try to prove it. Instead it might be more
fruitful to compute a basis of the cyclotomic integers and use
that to adjust the zoom levels, going back to uniform zoom for
continuous motion.

That could be a new kind of self-similarity.